backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
google-generativeai
pillow
python-magic
aiofiles
